Exclude test data files from crates.io package
