+++ src/euc_jp.rs
@@ -435,35 +435,4 @@ mod tests {
         encode_euc_jp("\u{3000}", b"\xA1\xA1");
         encode_euc_jp("\u{FF02}", b"\xFC\xFE");
     }
-
-    #[test]
-    #[cfg_attr(miri, ignore)] // Miri is too slow
-    fn test_jis0208_decode_all() {
-        let input = include_bytes!("test_data/jis0208_in.txt");
-        let expectation = include_str!("test_data/jis0208_in_ref.txt");
-        let (cow, had_errors) = EUC_JP.decode_without_bom_handling(input);
-        assert!(had_errors, "Should have had errors.");
-        assert_eq!(&cow[..], expectation);
-    }
-
-    #[test]
-    #[cfg_attr(miri, ignore)] // Miri is too slow
-    fn test_jis0208_encode_all() {
-        let input = include_str!("test_data/jis0208_out.txt");
-        let expectation = include_bytes!("test_data/jis0208_out_ref.txt");
-        let (cow, encoding, had_errors) = EUC_JP.encode(input);
-        assert!(!had_errors, "Should not have had errors.");
-        assert_eq!(encoding, EUC_JP);
-        assert_eq!(&cow[..], &expectation[..]);
-    }
-
-    #[test]
-    #[cfg_attr(miri, ignore)] // Miri is too slow
-    fn test_jis0212_decode_all() {
-        let input = include_bytes!("test_data/jis0212_in.txt");
-        let expectation = include_str!("test_data/jis0212_in_ref.txt");
-        let (cow, had_errors) = EUC_JP.decode_without_bom_handling(input);
-        assert!(had_errors, "Should have had errors.");
-        assert_eq!(&cow[..], expectation);
-    }
 }
