+++ tests/test_data/euc_kr_in.txt
