+++ tests/test_data/euc_kr_out.txt
