+++ tests/test_data/gb18030_in.txt
