+++ tests/test_data/gb18030_in_ref.txt
