+++ tests/test_data/gb18030_out.txt
