+++ tests/test_data/gb18030_out_ref.txt
