+++ tests/test_data/jis0212_in_ref.txt
