+++ tests/test_data/shift_jis_in_ref.txt
