+++ tests/test_data/shift_jis_out.txt
